Allow pings of different intervals
Since each check can now have its own polling interval, each failed poll
knows the duration since last poll, and the total downtime can be
computed in the MapReduce function.

Therefore the dashboard was a little reworked to add this information.

In addition, the erroneous polls now return a text error that is
displayed in the list of pings.