Fix race condition in api causing bad counts in dashboard.
A CheckEvent triggers both a refresh of the count in the api and in the
dashboard - the dashboard asking the api for a fresher value. When using
websockets, the response from MongoDB could happen after the request
from the dashboard.

The fix forces the dashboard to wait if a refresh was just triggered.