Update lifecycle event names.
The 'save' event being emitted before the actual persistence, it may happen that it gets emitted even if the underlying document doesn't get saved. Therefore it is necessary to create pre- and post- events.

This implies renaming all listeners, in the applications as well as in the plugins.

See https://groups.google.com/d/msg/mongoose-orm/UB5fwlIglok/53GER1voodEJ for details.