Added DNS protocoll
This poller checks the response time and health of a DNS server 
by polling it with an actual DNS query.
It can also be used to monitor the hostname resolution health of a single domain name. I might add the ability to test the DNS response from the server against an expected value (A-record).

URL format:

>dns://www.example.com[@8.8.8.8]

The domain name is required, the second parameter (IP address of DNS server) is optional and defaults to Google's public DNS server.